css/css-display/parsing/display-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Display: parsing display with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css2/visuren.html#display-prop">
<link rel="help" href="https://drafts.csswg.org/css-display/#the-display-properties">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#grid-containers">
<meta name="assert" content="display supports the full spec grammar.">
<meta name="assert" content="display supports the new values 'grid | inline-grid'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// https://drafts.csswg.org/css-grid-1/#grid-containers
test_valid_value("display", "grid");
test_valid_value("display", "inline-grid");

// https://drafts.csswg.org/css2/visuren.html#display-prop
test_valid_value("display", "inline");
test_valid_value("display", "block");
test_valid_value("display", "list-item");
test_valid_value("display", "inline-block");
test_valid_value("display", "table");
test_valid_value("display", "inline-table");
test_valid_value("display", "table-row-group");
test_valid_value("display", "table-header-group");
test_valid_value("display", "table-footer-group");
test_valid_value("display", "table-row");
test_valid_value("display", "table-column-group");
test_valid_value("display", "table-column");
test_valid_value("display", "table-cell");
test_valid_value("display", "table-caption");
test_valid_value("display", "none");

// https://drafts.csswg.org/css-display/#the-display-properties
test_valid_value("display", "run-in");
test_valid_value("display", "flow");
test_valid_value("display", "flow-root");
test_valid_value("display", "ruby");

test_valid_value("display", "flow list-item");
test_valid_value("display", "list-item flow");
test_valid_value("display", "flow-root list-item");
test_valid_value("display", "list-item flow-root");

test_valid_value("display", "block flow");
test_valid_value("display", "flow block");
test_valid_value("display", "flow-root block");
test_valid_value("display", "block flow-root");
test_valid_value("display", "flex block");
test_valid_value("display", "block flex");
test_valid_value("display", "grid block");
test_valid_value("display", "block grid");
test_valid_value("display", "table block");
test_valid_value("display", "block table");
test_valid_value("display", "block ruby");
test_valid_value("display", "ruby block");
test_valid_value("display", "block list-item");
test_valid_value("display", "list-item block");
test_valid_value("display", "flow block list-item");
test_valid_value("display", "block flow list-item");
test_valid_value("display", "flow list-item block");
test_valid_value("display", "block list-item flow");
test_valid_value("display", "list-item block flow");
test_valid_value("display", "list-item flow block");
test_valid_value("display", "flow-root block list-item");
test_valid_value("display", "block flow-root list-item");
test_valid_value("display", "flow-root list-item block");
test_valid_value("display", "block list-item flow-root");
test_valid_value("display", "list-item block flow-root");
test_valid_value("display", "list-item flow-root block");

test_valid_value("display", "inline flow");
test_valid_value("display", "flow inline");
test_valid_value("display", "flow-root inline");
test_valid_value("display", "inline flow-root");
test_valid_value("display", "flex inline");
test_valid_value("display", "inline flex");
test_valid_value("display", "grid inline");
test_valid_value("display", "inline grid");
test_valid_value("display", "table inline");
test_valid_value("display", "inline table");
test_valid_value("display", "inline ruby");
test_valid_value("display", "ruby inline");
test_valid_value("display", "inline list-item");
test_valid_value("display", "list-item inline");
test_valid_value("display", "flow inline list-item");
test_valid_value("display", "inline flow list-item");
test_valid_value("display", "flow list-item inline");
test_valid_value("display", "inline list-item flow");
test_valid_value("display", "list-item inline flow");
test_valid_value("display", "list-item flow inline");
test_valid_value("display", "flow-root inline list-item");
test_valid_value("display", "inline flow-root list-item");
test_valid_value("display", "flow-root list-item inline");
test_valid_value("display", "inline list-item flow-root");
test_valid_value("display", "list-item inline flow-root");
test_valid_value("display", "list-item flow-root inline");

test_valid_value("display", "run-in flow");
test_valid_value("display", "flow run-in");
test_valid_value("display", "flow-root run-in");
test_valid_value("display", "run-in flow-root");
test_valid_value("display", "flex run-in");
test_valid_value("display", "run-in flex");
test_valid_value("display", "grid run-in");
test_valid_value("display", "run-in grid");
test_valid_value("display", "table run-in");
test_valid_value("display", "run-in table");
test_valid_value("display", "run-in ruby");
test_valid_value("display", "ruby run-in");
test_valid_value("display", "run-in list-item");
test_valid_value("display", "list-item run-in");
test_valid_value("display", "flow run-in list-item");
test_valid_value("display", "run-in flow list-item");
test_valid_value("display", "flow list-item run-in");
test_valid_value("display", "run-in list-item flow");
test_valid_value("display", "list-item run-in flow");
test_valid_value("display", "list-item flow run-in");
test_valid_value("display", "flow-root run-in list-item");
test_valid_value("display", "run-in flow-root list-item");
test_valid_value("display", "flow-root list-item run-in");
test_valid_value("display", "run-in list-item flow-root");
test_valid_value("display", "list-item run-in flow-root");
test_valid_value("display", "list-item flow-root run-in");

</script>
</body>
</html>
